Fixed fatal bug with phone numbers not going past a certain amount. Turns out it was because I was using an Int to store the phone number which only goes up to (214).... Changed to long instead.

+++ src/main/kotlin/io/github/sgosiaco/library/model/Model.kt
@@ -33,7 +33,7 @@ class SelectedBook : ItemViewModel<Book>() {
 data class Person (
         @SerializedName("name") var name : String = "",
         @SerializedName("email") var email : String = "",
-        @SerializedName("phone") var phone : Int = -1,
+        @SerializedName("phone") var phone : Long = -1,
         @SerializedName("aff") var aff : String = "",
         @SerializedName("cNum") var cNum : Int = 0
 ) {
